Upgrade pytest to >= 3.6.0
There is no restriction on pytest-dist version, which introduces the
following error:

pytest-xdist 1.25.0 has requirement pytest>=3.6.0, but you'll have pytest 3.3.2 which is incompatible.

When upgrading to pytest >= 3.6.0, fixtures can no longer be called
directly as functions and thus the broken tests resulting from the
version bump of pytest had to be fixed.